TimelineItem refactor
