+++ src/components/Timeline/Timeline.module.scss
@@ -9,7 +9,6 @@
     .title {
         background: linear-gradient(45deg, white, #e0e7ff);
         background-clip: text;
-        background-clip: text;
         color: white;
         font-size: 2rem;
         font-weight: bold;
@@ -57,120 +56,6 @@
                 bottom: 0;
             }
         }
-
-        .timelineItem {
-            margin-bottom: 3rem;
-            position: relative;
-            width: 100%;
-
-            .timelineMarker {
-                left: calc(-3rem - 0.5 * var(--dot-width) + 0.5 * var(--line-width));
-                position: absolute;
-                top: calc(50% - 0.5 * var(--dot-width));
-                z-index: 2;
-
-                .markerDot {
-                    background: white;
-                    border: 3px solid rgb(white, 0.3);
-                    border-radius: 50%;
-                    box-shadow: 0 0 10px rgb(white, 0.3);
-                    height: var(--dot-width);
-                    transition: all 0.3s ease;
-                    width: var(--dot-width);
-                }
-            }
-
-            .timelineContent {
-                backdrop-filter: blur(10px);
-                background: rgb(white, 0.85);
-                border: 1px solid rgb(white, 0.2);
-                border-radius: 16px;
-                box-shadow: 0 8px 32px rgb(0 0 0 / 10%);
-                cursor: pointer;
-                opacity: 0;
-                padding: 1.5rem;
-                transform: translateX(50px);
-                transition: all 0.3s cubic-bezier(0.4, 0, 0.2, 1);
-                transition: all 0.6s;
-                width: 100%;
-
-                &.visible {
-                    opacity: 1;
-                    transform: translateX(0);
-                }
-
-                .contentHeader {
-                    margin-bottom: 1rem;
-
-                    .dateRange {
-                        align-items: center;
-                        display: flex;
-                        flex-wrap: wrap;
-                        font-size: 0.875rem;
-                        font-weight: 500;
-                        gap: 0.25rem;
-                        margin-bottom: 0.75rem;
-
-                        .startDate,
-                        .endDate {
-                            color: $primary-color;
-                            font-weight: 600;
-                        }
-
-                        .dateSeparator {
-                            color: $primary-color;
-                        }
-
-                        .duration {
-                            background: #f5ebf8;
-                            border-radius: 6px;
-                            color: $primary-color;
-                            font-size: 0.75rem;
-                            font-weight: 600;
-                            padding: 0.25rem 0.5rem;
-                        }
-                    }
-
-                    .position {
-                        color: $dark-bg-color;
-                        font-size: 1.25rem;
-                        font-weight: bold;
-                        line-height: 1.3;
-                        margin-bottom: 0.25rem;
-                    }
-
-                    .company {
-                        color: $primary-color;
-                        font-size: 1rem;
-                        font-weight: 600;
-                        margin-bottom: 0;
-                    }
-                }
-
-                .achievements {
-                    margin-bottom: 1rem;
-                    max-height: 300px;
-                    overflow: hidden;
-                    transition: all 0.4s cubic-bezier(0.4, 0, 0.2, 1);
-
-                    .achievementsList {
-                        color: $dark-bg-color;
-                        list-style: disc outside;
-
-                        .achievement {
-                            line-height: 1.5;
-                            margin-bottom: 0.5rem;
-                            margin-left: 2rem;
-                            position: relative;
-
-                            &:last-child {
-                                margin-bottom: 0;
-                            }
-                        }
-                    }
-                }
-            }
-        }
     }
 }
 
@@ -188,53 +73,6 @@
 
             .line {
                 left: calc(-0.5 * var(--dot-width) - 0.5 * var(--line-width));
-            }
-
-            .timelineItem {
-                .timelineMarker {
-                    left: -3rem;
-                }
-
-                .timelineContent {
-                    padding: 2rem;
-
-                    .contentHeader {
-                        .dateRange {
-                            flex-wrap: nowrap;
-                            gap: 0.5rem;
-                        }
-
-                        .position {
-                            font-size: 1.5rem;
-                        }
-
-                        .company {
-                            font-size: 1.125rem;
-                        }
-                    }
-                }
-            }
-        }
-    }
-}
-
-// Hover effects
-@media (hover: hover) and (pointer: fine) {
-    .timeline {
-        .timelineBody {
-            .timelineItem {
-                .timelineMarker {
-                    .markerDot:hover {
-                        box-shadow: 0 0 15px rgb(white, 0.5);
-                        transform: scale(1.1);
-                    }
-                }
-
-                .timelineContent:hover {
-                    background: rgb(white, 0.9);
-                    box-shadow: 0 12px 40px rgb(0 0 0 / 15%);
-                    transform: translateY(-2px);
-                }
             }
         }
     }
