+++ src/components/Timeline/Timeline.tsx
@@ -1,19 +1,10 @@
 // components/Timeline.tsx
 'use client';
 
+import { WorkExperience } from '@/utils/interfaces';
 import { useEffect, useRef, useState } from 'react';
+import TimelineItem from '../TimelineItem/TimelineItem';
 import styles from './Timeline.module.scss';
-
-interface WorkExperience {
-    id: string;
-    company: string;
-    position: string;
-    startDate: string;
-    endDate: string;
-    description: string;
-    achievements: string[];
-    logo?: string;
-}
 
 const workExperiences: WorkExperience[] = [
     {
@@ -99,37 +90,6 @@ const Timeline: React.FC = () => {
         return () => observer.disconnect();
     }, []);
 
-    // Helpers
-    const formatDate = (dateString: string): string => {
-        if (dateString === 'Present') return 'Present';
-        const date = new Date(dateString);
-        return date.toLocaleDateString('en-US', {
-            year: 'numeric',
-            month: 'short',
-        });
-    };
-
-    const calculateDuration = (start: string, end: string): string => {
-        const startDate = new Date(start);
-        const endDate = end === 'Present' ? new Date() : new Date(end);
-        const diffTime = Math.abs(endDate.getTime() - startDate.getTime());
-        const diffMonths = Math.ceil(diffTime / (1000 * 60 * 60 * 24 * 30));
-
-        if (diffMonths < 12) {
-            return `${diffMonths} month${diffMonths !== 1 ? 's' : ''}`;
-        }
-
-        const years = Math.floor(diffMonths / 12);
-        const months = diffMonths % 12;
-
-        let duration = `${years} year${years !== 1 ? 's' : ''}`;
-        if (months > 0) {
-            duration += ` ${months} month${months !== 1 ? 's' : ''}`;
-        }
-
-        return duration;
-    };
-
     // JSX
     return (
         <div className={styles.timeline}>
@@ -137,51 +97,11 @@ const Timeline: React.FC = () => {
             <div className={styles.timelineBody} ref={timelineRef}>
                 <div className={styles.line} />
                 {workExperiences.map(experience => (
-                    <div
+                    <TimelineItem
                         key={experience.id}
-                        data-id={experience.id}
-                        className={styles.timelineItem}
-                    >
-                        <div className={styles.timelineMarker}>
-                            <div className={styles.markerDot} />
-                        </div>
-                        <div
-                            className={`${styles.timelineContent}  ${
-                                visibleItems.has(experience.id) ? styles.visible : ''
-                            }`}
-                        >
-                            <div className={styles.contentHeader}>
-                                <div className={styles.dateRange}>
-                                    <span className={styles.startDate}>
-                                        {formatDate(experience.startDate)}
-                                    </span>
-                                    <span className={styles.dateSeparator}>—</span>
-                                    <span className={styles.endDate}>
-                                        {formatDate(experience.endDate)}
-                                    </span>
-                                    <span className={styles.duration}>
-                                        (
-                                        {calculateDuration(
-                                            experience.startDate,
-                                            experience.endDate
-                                        )}
-                                        )
-                                    </span>
-                                </div>
-                                <h3 className={styles.position}>{experience.position}</h3>
-                                <h4 className={styles.company}>{experience.company}</h4>
-                            </div>
-                            <div className={styles.achievements}>
-                                <ul className={styles.achievementsList}>
-                                    {experience.achievements.map((achievement, achIndex) => (
-                                        <li key={achIndex} className={styles.achievement}>
-                                            {achievement}
-                                        </li>
-                                    ))}
-                                </ul>
-                            </div>
-                        </div>
-                    </div>
+                        experience={experience}
+                        visible={visibleItems.has(experience.id)}
+                    />
                 ))}
             </div>
         </div>
